pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "A sample MCP server implementation"
requires-python = ">=3.8"
dependencies = [
    "mcp[cli]",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "requests",
    "httplib2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]

[project.scripts]
mcp-server = "mcp_server.server:main"
